update biglib too

+++ crab/multiarch/run.sh
@@ -1,14 +1,18 @@
-#!/bin/bash -ex
+#!/bin/bash -e
 pwd
 ls
 env > run.log
 pushd $CMSSW_BASE
-  rm -rf lib
-  rsync -a $CMSSW_RELEASE_BASE/lib/ ./lib/ || true
-  cd lib/$SCRAM_ARCH/scram_x86-64-v2
-  for pcm in $(ls *.pcm) ; do
-    rm -f $pcm
-    cp ../$pcm .
+  for dir in lib biglib ; do
+    rm -rf $dir
+    rsync -a $CMSSW_RELEASE_BASE/${dir}/ ./${dir}/ || true
+    [ -d ${dir}/$SCRAM_ARCH/scram_x86-64-v2 ] || continue
+    pushd ${dir}/$SCRAM_ARCH/scram_x86-64-v2
+      for pcm in $(ls *.pcm) ; do
+        rm -f $pcm
+        cp ../$pcm .
+      done
+    popd
   done
 pushd
 export LD_LIBRARY_PATH=$(echo $LD_LIBRARY_PATH | tr : '\n' | grep -E -v "$CMSSW_RELEASE_BASE/(big|)lib/" | grep -E -v "/${CMSSW_VERSION}/(big|)lib/${SCRAM_ARCH}$" | tr '\n' ':')
